Simple changes to support user-supplied league data
The www.pathofexile.com/trade site 403s apparently due to CloudFlare spam protection (see #8384).

If you're using POESESSID, this means you can't use the trade tool until you null-out the POESESSID, to use the static league data.

This branch attempts to fall back to a new field that can be user-supplied, Misc->TradeLeagueData.

This is intended to be grabbed from the source of www.pathofexile.com/trade by the user, rather than POB trying to grab it.

If that doesn't exist, the tool will finally fall back to the static list.

A UI entry isn't provided in this branch.